+++ app/_components/footer.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../styles/scss/Footer.module.scss";
+import styles from "../_styles/scss/Footer.module.scss";
 const Footer = () => {
   return <div className={styles.footer}>I am footer</div>;
 };
