+++ app/_styles/scss/Footer.module.scss
