+++ app/_styles/scss/Header.module.scss
@@ -1,0 +1,21 @@
+.header {
+  // todo
+  background-color: deeppink;
+
+  .headerLogoWrapper {
+    // todo
+
+    .headerLogo {
+      // todo
+    }
+    .headerLogoText {
+      // todo
+    }
+    .headerNavigation {
+      // todo
+    }
+    .headerAccount {
+      // todo
+    }
+  }
+}
