+++ app/about/page.tsx
@@ -3,7 +3,7 @@ import React from "react";
 const About = () => {
   return (
     <div>
-      <h1 className="text-blue-600">About</h1>
+      <h1>About</h1>
     </div>
   );
 };
