+++ app/blog/page.tsx
@@ -3,7 +3,7 @@ import React from "react";
 const Blog = () => {
   return (
     <div>
-      <h1 className="text-blue-600">Blog</h1>
+      <h1>Blog</h1>
     </div>
   );
 };
