+++ app/page.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Page() {
+  return <div>I am home page</div>;
+}
