+++ app/schedule/page.tsx
@@ -3,7 +3,7 @@ import React from "react";
 const Schedule = () => {
   return (
     <div>
-      <h1 className="text-blue-600">Schedule</h1>
+      <h1>Schedule</h1>
     </div>
   );
 };
