--- src/app/dashboard/layout.tsx
@@ -1,18 +1,0 @@
-import { Metadata } from "next";
-import Page from "./(overview)/page";
-// import Header from './header/page';
-
-export const metadata: Metadata = {
-  title: "Invoices",
-};
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return (
-    <div className="flex h-screen flex-col md:flex-row md:overflow-hidden">
-      <div className="w-full flex-none md:w-64">
-        <Page />
-        <h1>HI</h1>
-      </div>
-      <div className="flex-grow p-6 md:overflow-y-auto md:p-12">{children}</div>
-    </div>
-  );
-}
