--- src/app/styles/scss/Header.module.scss
@@ -1,3 +1,0 @@
-.header {
-  background-color: deeppink;
-}
